requirements.txt
gradio
requests
google-genai
dotenv
pytest